Ejercicio revisado. Cambiado int a double

+++ Ejercicio_4_1_1_1/Ejercicio_4_1_1_1/Ejercicio_4_1_1_1/Program.cs
@@ -6,22 +6,22 @@ namespace Ejercicio_4_1_1_1
     {
         static void Main(string[] args)
         {
-            int[] numeros = new int[4];
-            int suma;
-            int division;
+            double[] numeros = new double[4];
+            double suma;
+            double division;
             
 
             Console.WriteLine("Introduce el primer número: ");
-            numeros[0] = Convert.ToInt32(Console.ReadLine());
+            numeros[0] = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine("Introduce el segundo número: ");
-            numeros[1] = Convert.ToInt32(Console.ReadLine());
+            numeros[1] = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine("Introduce el tercer número: ");
-            numeros[2] = Convert.ToInt32(Console.ReadLine());
+            numeros[2] = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine("Introduce el cuarto número: ");
-            numeros[3] = Convert.ToInt32(Console.ReadLine());
+            numeros[3] = Convert.ToDouble(Console.ReadLine());
 
             suma = numeros[0] + numeros[1] + numeros[2] + numeros[3];
             division = suma / 4;
